requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.1
groq>=1.40.0

# Testing
pytest>=9.0.0
anyio>=4.0.0
pytest_asyncio>=1.3.0
